Version 2 Improvements
Update to ES6
Applied changes according to following feedback:
We noticed that your solution assumed that the ids in the files are incremented by one and are in order. We had test cases where this was not the case, and your solution could not handle those cases. Secondly, the method of calculating total average at the moment is incorrect.
We like how you used the index of the array to make the solution more efficient, however, as you can see, it fails other test cases. Instead, consider storing the data in objects (i.e. hash tables), so you can efficiently reference information by id, but doesn't make any assumptions about those ids.